restructured import.go and everything related to it. import_test is left.
