+++ internal/commands/util/completion.go
@@ -6,6 +6,7 @@ import (
 	"os"
 
 	"github.com/MakeNowJust/heredoc"
+	"github.com/checkmarx/ast-cli/internal/shared"
 	"github.com/spf13/cobra"
 )
 
@@ -62,7 +63,7 @@ func NewCompletionCommand() *cobra.Command {
 		Args: func(cmd *cobra.Command, args []string) error {
 			shellType, _ := cmd.Flags().GetString(shellFlag)
 
-			if shellType == "" || Contains(cmd.ValidArgs, shellType) {
+			if shellType == "" || shared.Contains(cmd.ValidArgs, shellType) {
 				return nil
 			}
