+++ internal/commands/util/import.go
@@ -1,4 +1,4 @@
-package commands
+package util
 
 import (
 	"path/filepath"
@@ -8,6 +8,7 @@ import (
 	"github.com/checkmarx/ast-cli/internal/constants"
 	errorConstants "github.com/checkmarx/ast-cli/internal/constants/errors"
 	commonParams "github.com/checkmarx/ast-cli/internal/params"
+	"github.com/checkmarx/ast-cli/internal/shared"
 	"github.com/checkmarx/ast-cli/internal/wrappers"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
@@ -59,7 +60,7 @@ func runImportCommand(
 			return errors.Errorf(errorConstants.ProjectNameIsRequired)
 		}
 
-		projectID, err := findProject(nil, projectName, cmd, projectsWrapper, groupsWrapper, accessManagementWrapper)
+		projectID, err := shared.FindProject(nil, projectName, cmd, projectsWrapper, groupsWrapper, accessManagementWrapper)
 		if err != nil {
 			return err
 		}
